requirements.txt
rich>=10.16.2
click>=8.1.3
httpx>=0.24.1
tenacity>=8.2.2
chromadb>=0.4.13
pdfminer.six>=20221105
pyyaml>=6.0
asyncio>=3.4.3
markdown>=3.4.3
tiktoken>=0.4.0
pytest>=7.3.1
pytest-cov>=4.1.0
pre-commit>=3.3.2
flake8>=6.0.0
black>=23.3.0
mypy>=1.3.0
python-dotenv>=1.0.0